+++ client/src/Components/Header.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { NavLink } from 'react-router-dom'
 import { useState, useContext } from 'react'
-import LoginContext from './SignInContext'
+// import LoginContext from './SignInContext'
 
 
 
@@ -13,8 +13,8 @@ export default function Header() {
             <NavLink to="/about">About</NavLink>
             <NavLink to="/explore">Explore</NavLink>
             <NavLink to="/">Good Hikes</NavLink>
-            <NavLink to="/login">{LoginContext ? "Logout" : "Login"}</NavLink>
-            <NavLink to="/signup">{LoginContext ? "Account" : "Sign up"}</NavLink>
+            <NavLink to="/login">Login</NavLink>
+            <NavLink to="/signup">Signup</NavLink>
             
         </header>
     )
