+++ client/src/Components/Login.js
@@ -1,71 +1,29 @@
 import React from "react";
 import { useState, createContext, useContext } from "react"
 import { useNavigate } from "react-router-dom";
-import LoginContext from "./SignInContext";
+// import LoginContext from "./SignInContext";
+import LoginForm from "./LoginForm";
 
 
 export default function Login() {
 
   // const signedIn = useContext(isSignedIn)
 
-  console.log(LoginContext == true)
+  // console.log(LoginContext == true)
   
-  const [errors, setErrors] = useState([])
-  const [formData, setformData] = useState({
-    username: "",
-    password: "",
-  });
-
-  const navigate = useNavigate()
-  const { username, password } = formData
-
   
-  function handleChange(e) {
-    setformData({
-      ...formData,
-      [e.target.name]: e.target.value
-    })
-  }
-
-  //double check this
-  function handleSubmit(e) {
-    e.preventDefault()
-    const user = {
-      username, password
-    }
-    console.log(user)
-    fetch('/login', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(user)
-    })
-      .then(res => res.json())
-      .then(data=>console.log(data))
-      //change signed in context here, then redirect to explore page
-      .then(navigate('/explore'))
-      .catch(error => setErrors(error))
-      .then(console.log("errors", errors))
-  }
 
   
   return (
+    
     <main className="login-page">
       
-      <div className="form-div">
-        <form onSubmit={handleSubmit}>
-          
-          <label>Username</label>
-          <input type="text" name="username" value={username} onChange={handleChange} />
-          
-          <label>Password</label>
-          <input type="password" name="password" value={password} onChange={handleChange} />
-          <button type="submit">Login</button>
-        
-        </form>
-      </div>
+      <LoginForm />
+      
       {/* {signedIn? "TEST you are signed in" : " TEST signin now"} */}
-      {errors ? errors.map(e => <div>{e[0] + ': ' + e[1]}</div>) : null}
+      {/* {errors ? errors.map(e => <div>{e[0] + ': ' + e[1]}</div>) : null} */}
     </main>
+    
     //display map over errors array
     //set timeout to display errors, then reset state back to empty array, trigger rerender
